ロジックファイルとビューファイルの分離化 プロフィール画面と投稿フォームのみ対応＋muiのインポートの共通化
