+++ src/muiImports.js
@@ -1,0 +1,13 @@
+// マテリアルUIの共通化
+// UI
+export { default as Button } from '@mui/material/Button';
+export { default as TextField } from '@mui/material/TextField';
+export { default as Fab } from '@mui/material/Fab';
+export { default as Dialog } from '@mui/material/Dialog';
+export { default as DialogActions } from '@mui/material/DialogActions';
+export { default as DialogContent } from '@mui/material/DialogContent';
+export { default as DialogContentText } from '@mui/material/DialogContentText';
+export { default as DialogTitle } from '@mui/material/DialogTitle';
+// アイコン
+export { default as EditIcon } from '@mui/icons-material/Edit';
+export { default as SendIcon } from '@mui/icons-material/Send';
